Change type of ARRAY_COUNT() to unsigned long to avoid problems on 32-bit systems
On 64-bit systems size_t is unsigned long but on 32-bit systems it is
unsigned int. That means you can't use neither %lu nor %u printf format
specifier. You could use %z but that's C99.

+++ common.h
@@ -28,7 +28,7 @@
  * argument at index @first_idx is the first format argument */
 #define __FORMAT(fmt_idx, first_idx) __attribute__((format(printf, (fmt_idx), (first_idx))))
 
-#define ARRAY_COUNT(x) (sizeof(x) / sizeof(x[0]))
+#define ARRAY_COUNT(x) ((unsigned long)sizeof(x) / sizeof(x[0]))
 
 #define clear(ptr) memset((ptr), 0, sizeof(*(ptr)))
